+++ src/rics/__main__.py
@@ -1,6 +1,6 @@
 """Alias of :mod:`rics.cli`.
 
-As per, https://docs.python.org/3/library/__main__.html#id1,
+As per, https://docs.python.org/3.13/library/__main__.html#id1,
 this file doesn't use a ``if __name__ == "__main__"``-block.
 """