+++ src/rics/envinterp/__init__.py
@@ -17,3 +17,6 @@ if os.environ.get("SPHINX_BUILD") == "true":
     # Ugly fix for duplicate indexes.
     del __all__, UnsetVariableError, Variable, replace_in_string
 del os
+
+# TODO(6.0.0): Deprecate this module
+# TODO(7.0.0): Remove this module