+++ src/rics/performance/__init__.py
@@ -1,15 +1,18 @@
 """Performance testing utility."""
 
 from ._format_perf_counter import format_perf_counter, format_seconds
-from ._multi_case_timer import MultiCaseTimer
-from ._util import get_best, plot_run, to_dataframe
+from ._multi_case_timer import MultiCaseTimer, SkipIfParams
+from ._plot import plot_run
+from ._util import get_best, legacy_plot_run, to_dataframe
 from ._wrapper import run_multivariate_test
 
 __all__ = [
     "MultiCaseTimer",
+    "SkipIfParams",
     "format_perf_counter",
     "format_seconds",
     "get_best",
+    "legacy_plot_run",
     "plot_run",
     "run_multivariate_test",
     "to_dataframe",