+++ tests/performance/test_plot_module/test_plot.py
@@ -3,21 +3,20 @@ import json
 import shutil
 from functools import cache
 from pathlib import Path
-from typing import Any, cast, get_args
+from typing import Any, cast
 
 import pytest
 from matplotlib import pyplot as plt
 
+from rics.performance import plot_run
+from rics.performance.plot_types import Kind
 from rics.performance.types import ResultsDict
-
-with pytest.warns(UserWarning, match="The rics.performance.plot API is experimental."):
-    from rics.performance.plot import plot
-    from rics.performance.plot.types import Kind
 
 pytestmark = [
     pytest.mark.filterwarnings("ignore:.*points cannot be placed:UserWarning"),
     pytest.mark.filterwarnings("ignore:The test results may be unreliable:UserWarning"),
-    pytest.mark.parametrize("kind", get_args(Kind)),
+    # pytest.mark.parametrize("kind", get_args(Kind)),  # For some reason, 'swarm' is very slow.
+    pytest.mark.parametrize("kind", ["bar", "violin"]),
 ]
 
 OUT_ROOT = Path(__file__).parent / "test-output/plots"
@@ -26,18 +25,13 @@ with contextlib.suppress(FileNotFoundError):
 
 
 def run(kind: Kind, **kwargs: Any) -> None:
-    root = OUT_ROOT.joinpath(*kwargs) if kwargs else OUT_ROOT.joinpath("default")
-    name = "-".join([kind, *kwargs.values()]) + ".jpg"
-    path = root / name
-    assert not path.exists()
-
-    results = run_results()
+    run_results = get_run_results()
     if kind == "box":
         # seaborn==0.13.2 / matplotlib==3.10.0
         with pytest.warns(PendingDeprecationWarning, match="vert: bool will be deprecated"):
-            facet_grid = plot(results, kind=kind, path=path, **kwargs)
+            facet_grid = plot_run(run_results, kind=kind, **kwargs)
     else:
-        facet_grid = plot(results, kind=kind, path=path, **kwargs)
+        facet_grid = plot_run(run_results, kind=kind, **kwargs)
 
     plt.close(facet_grid.fig)
 
@@ -52,7 +46,7 @@ def test_estimator(kind: Kind, estimator: str) -> None:
 
 
 @cache
-def run_results() -> ResultsDict:
+def get_run_results() -> ResultsDict:
     with Path(__file__).parent.joinpath("run-results.json").open("r") as f:
         results = json.load(f)
     return cast(ResultsDict, results)