+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * Added `types.LiteralHelper[T].read_env()`.
 - New function `strings.str_as_bool()`.
 - New module `rics.click`.
+- The `MultiCaseTimer` now supports callable `test_data` args to generate data.
+- Finished new `rics.performance` plotting implementation (support facets).
 
 ### Changed
 - Update `basic_config.basic_config()`: Allow and handle `level=None` to avoid logging from root.